servicebusapi/requirements.txt
fastapi 
pandas
azure-servicebus
uvicorn[standard]
pyodbc